[All] add missing type declarations and declare_strict declaration

+++ src/CoreShop/Bundle/CoreBundle/Validator/Constraints/CartMaximumQuantityValidator.php
@@ -98,7 +98,7 @@ final class CartMaximumQuantityValidator extends ConstraintValidator
         }
     }
 
-    private function getExistingCartItemQuantityFromCart(OrderInterface $cart, OrderItemInterface $cartItem): int
+    private function getExistingCartItemQuantityFromCart(OrderInterface $cart, OrderItemInterface $cartItem): float
     {
         $product = $cartItem->getProduct();
         $quantity = $cartItem->getDefaultUnitQuantity();
